refactor(models): move flow greeks and options contracts
